Started static page to search people and started search of people in page to create group
Added the page with a partial to search people and a redirection to that page from the search_posts page.
The query in UserController needs fixing.
The search in the page that creates a group will be a headache, need to make rows properly without hiding necessary users/showing unnecessary ones.
Also completed Xico's user story of this week